+++ web/src/modules/PrivateView/index.tsx
@@ -82,6 +82,9 @@ const PagePrivateView = () => {
             }}
           />
         )}
+
+        {/* Extra scripts */}
+        <script>const a = 123;</script>
       </Flex>
     </>
   );